create cloneProdToDev db helper, fix undefined self error
