+++ src/server/database/cloneProdToDev.babel.js
@@ -1,0 +1,2 @@
+require('babel-register');
+require('./cloneProdToDev').default();
