+++ src/server/graphql/models/Action/updateAction/updateActionValidation.js
@@ -7,22 +7,23 @@ export default function updateActionValidation(tms) {
         const [teamId] = id.split('::');
         return !tms.includes(teamId) && `Not on team ${teamId}`;
       }),
-    agendaId: (value) => value
-      .test((id) => {
-        if (!id) return undefined;
-        const [teamId] = id.split('::');
-        return !tms.includes(teamId) && `Not on team ${teamId}`;
-      }),
+    // agendaId: (value) => value
+    //   .test((id) => {
+    //     if (!id) return undefined;
+    //     const [teamId] = id.split('::');
+    //     return !tms.includes(teamId) && `Not on team ${teamId}`;
+    //   }),
     content: (value) => value
       .trim()
       .max(255, 'Whoa! That looks like 2 actions'),
     isComplete: (value) => value.boolean(),
     sortOrder: (value) => value.float(),
-    // teamMemberId: (value) => value
-    //   .test((id) => {
-    //     if (!id) return undefined;
-    //     const [, teamId] = id.split('::');
-    //     return !tms.includes(teamId) && `Not on team ${teamId}`;
-    //   })
+    // required for actions created in meetings
+    teamMemberId: (value) => value
+      .test((id) => {
+        if (!id) return undefined;
+        const [, teamId] = id.split('::');
+        return !tms.includes(teamId) && `Not on team ${teamId}`;
+      })
   });
 }
