+++ src/universal/utils/sendAssetToS3.js
@@ -1,6 +1,7 @@
-import fetch from 'universal/utils/fetch';
+// import fetch from 'universal/utils/fetch';
 
 export default async function sendAssetToS3(asset, assetPutUrl) {
+  if (!__CLIENT__) return undefined;
   const res = await fetch(assetPutUrl, {
     method: 'PUT',
     body: asset
